Add migrations with Liquibase and improve data structure in the database

+++ fsms-backend-api/src/main/java/com/fuelstation/managmentapi/fuelorder/domain/FuelOrderFactory.java
@@ -2,6 +2,8 @@ package com.fuelstation.managmentapi.fuelorder.domain;
 
 import com.fuelstation.managmentapi.common.domain.FuelGrade;
 
+import java.math.BigDecimal;
+
 public interface FuelOrderFactory {
-    FuelOrder create(Long fuelStationId, FuelGrade fuelGrade, float amount);
+    FuelOrder create(Long fuelStationId, FuelGrade fuelGrade, BigDecimal amount);
 }
