+++ fsms-backend-api/src/main/java/com/fuelstation/managmentapi/fuelstation/domain/exceptions/TankCapacityExceededException.java
@@ -2,16 +2,18 @@ package com.fuelstation.managmentapi.fuelstation.domain.exceptions;
 
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 /**
  * Exception thrown when a refill would exceed tank capacity
  */
 @Getter
 public class TankCapacityExceededException extends FuelStationDomainException {
     private final long fuelTankId;
-    private final float requestedVolume;
-    private final float maxCapacity;
+    private final BigDecimal requestedVolume;
+    private final BigDecimal maxCapacity;
 
-    public TankCapacityExceededException(long fuelTankId, float requestedVolume, float maxCapacity) {
+    public TankCapacityExceededException(long fuelTankId, BigDecimal requestedVolume, BigDecimal maxCapacity) {
         super(
             String.format("Refill request of %.2f liters exceeds tank %d capacity of %.2f liters", 
                 requestedVolume, fuelTankId, maxCapacity),
