+++ fsms-backend-api/src/main/java/com/fuelstation/managmentapi/fuelstation/domain/models/FuelPrice.java
@@ -2,4 +2,6 @@ package com.fuelstation.managmentapi.fuelstation.domain.models;
 
 import com.fuelstation.managmentapi.common.domain.FuelGrade;
 
-public record FuelPrice(FuelGrade fuelGrade, float pricePerLiter) {}
+import java.math.BigDecimal;
+
+public record FuelPrice(FuelGrade fuelGrade, BigDecimal pricePerLiter) {}
